Update NFTLendAuctionV1.sol
Fixes to audit findings:


* CEN-01Centralized and controlled privileges

We'll transfer ownership to a Safe multi-sig after deployment.

* CEN-03: Lack of circuit breaker

We do have have a maxActiveLoans parameter to circuit break the creation of new listed loans in case is needed


* LOG-02: Potential front-running

This is a reversed auction market, and the behaviour of placing better bids is intended.


* COD-03: Potential denial of service (DoS)

Added bid cooldown period of 1 hour
  ... bidTimestamps
  ... bidCancelPeriod


* LOG-04-01: Checks-Effects-Interactions

Moved transfers to be done _after_ all state changes, and added a try/catch block to all external calls with a revert.

* COD-01: Dependency on NFT contracts

We have a governance parameter called "allowedNFTContracts" to control which Collections are vetted for use.


* COD-11: Insufficient validation of loan conditions

a) we have a "onlyNftOwner" modifier that checks NFT ownership before listing the loan. Once is listed, the owner will be our contract.

b) "Interest rate calculations in getTotalRepayment are straightforward but do not account for complex scenarios such as early repayments or late payments beyond the simple fixed APR."

We do have a fallback to minimun "1 day" if the repay time is less than a day:
 
... (elapsedTimeInDays > 0 ? elapsedTimeInDays : 1)...


c) Added to acceptLoan

"require(escrowedFunds[loanId] == loan.loanAmount, "Escrowed funds do not match loan amount");"


* COD-12Lack of event-driven architecture

Added MaxActiveLoansUpdated event (setProtocolFeeRate already had ProtocolFeeRateUpdated)

* COM-01: Floating pragma
Changed to be exactly 0.8.23


* COM-03: Protocol fee calculation transparency

Added function calculateProtocolFee()

* COM-04Gas optimization

We already have a maxActiveLoans limit for the array to prevent this.